Fix plugins after l4d2 update
Fix of offsets after the update, now it will dynamically find new offsets if possible, the rest will be in the gamedata 'l4d2_si_ability', and updating the syntax. It was hard to find it and fix it, and these are not all plugins requiring updating. Give me chocolate for that =))