Restore default chat behavior while paused
The problem is that `gpGlobals->curtime` doesn't change in pause, it's used by dataprop `m_fLastPlayerTalkTime`,
so chat doesn't work during pause, previously the code in the plugin compensated for this,
which changed the default game behavior, and all the code in the game function `Host_Say` was ignored, and some things,
such as plugins intercepting the `player_say` event, did not work correctly.

In fact, in this function `Host_Say`, there is a lot of code, much more than in plugin `pause`,
for example adding a prefix to the message,
formatting the text in the correct form,
checking the validity of the message, creating event `player_say`,
checking whether each player can hear you, etc.

Some bugs for example:
`HlstastsX` (used to collect stats and chat) did not collect chat stats.
Plugins allowing to ignore another player could not work.